Allow specifying a revision to stop at using the -H flag.
This is useful for converting repositories which have been deleted or
renamed, such as llvm-gcc-4-2 in the LLVM repositories which was
renamed to llvm-gcc-4.2 shortly after its creation.

Also, consolidate the two places in svn_swig_wrapper.py where a
default chunk size is specified to one, single variable declaration.