svnmeta: improve error message in get_source_rev()
Letting a KeyError propagate when the supplied revision has no
conversion record on file isn't terribly helpful. At least not when
this KeyError merely states that the key missing is
'convert_revision'.  It's much better to preemptively check for it,
detect that it's missing, and then raise an exception with a
descriptive message.