help topic: cosmetic improvements.
First, use of :hg:`...` is replaced with ``hg ...``. The former syntax
isn't useful outside core Mercurial. Second, a few instances of `...`
are replaced with ``...``. The minirst parser doesn't distinguish
between the two, but using docutils, the former results italics and
the latter in fixed width text. Finally, a few extra ``...`` are added.

With these changes, we could process the help topic with rst2html and
put it somewhere appropriate on the internet.