stupid/filemap: disable this since it doesn't currently work
hg 1.9 dramatically cleaned up patch application, but unfortunately
this breaks stupid mode with filemaps. In the name of getting a
release out the door, disabling this feature for now. There will be
changes required in hg to make this work again, so we may just drop
the feature entirely if nobody's interested.