editor: work around svn 1.7 closing the output fp in apply_textdelta()
Calling close() on the output cStringIO when done with patching prevents us to
retrieve the result with getvalue().

Note: I found this bug has been fixed already by 794f3e41ec1e. I feel ignoring
the close() call is a little cleaner than storing the value before the
handler() call for there is no guarantee the handler does not modify the value
in its last call. Looking at text_delta.c, it does not in 1.7.1. So, do what
you want with this patch.