editor: resolve files copied by directory copy on-demand
Before this change, the data of files brought by directory copies was
resolved immediately and stored for the duration of the run. Now, only
references are stored and are resolved either when opening the files or
when closing the editor. It means RevisionData file data is only set
once per file and only when the file has been handled.

The next step is to turn RevisionData into a data store backed by the
filesystem.