pull: save the last pulled revision *after* converting revisions
Otherwise, a KeyboardInterrupt may lead to an unpulled revision being
incorrectly saved as pulled in the lastpulled file. This will lead to
the interrupted revision being incorrectly skipped at the next pull,
leading to an incorrect conversion -- one might even say corrupt.

Due to it's nature of requiring a manual interrupt, this bug is
difficult to test.