buildmeta: verify uuid when passed explicit url
2c87bdc43d3c changed buildmeta to read the uuid from local disk, if
it's available.  As a side effect, it disabled the validation of the
uuid we've recorded locall in the commits against the uuid of the repo
we are building against.  This is probably reasonable when someone
runs a bare updatemeta or rebuildmeta.  I think it's a worthwhile
sanity check when someone passes an explicit repository url to
rebuild/updatemeta.  This restores the validation in that case, and
fixes a failing test in the process.