commit: fix exception handling on transaction close
On SVN 1.6 at least, a failing pre-commit hook appears to abort a
commit transaction without us needing to do it manually. This is
apparently managed by a callback at the end of SVN's internal commit
callback chain.

Subsequently trying to abort an already-aborted transaction results in
a spurious "Transaction cleanup failed" message that breaks tests.

To fix this, we simply move the commit call out of try/except block.