layouts: consistently return None for default branch
This makes the single and standard layouts consistent in what they
return for the default branch.  Previously, single had returned
'default' now they both return None.

In addition, this fixes a now-exposed bug in stupid's convert_revision
logic.  Specifically, when a file is replaced by another file within
the same branch, we treated that as replacing the entire branch.  this
bug was previously hidden because meta.split_branch_path and
meta.localname were inconsistent in what they returned for the single
layout.  meta.split-branch_path is used to maintain the set of known
branches, where meta.localname is used to determine the branch for the
path being replaced.  This resulted in erroneously hitting the
condition that skipped paths outside branches we know about when
considering replace operations from svn.