svnwrap: use get_file instead of list_props during replay
This patch has an enormous effect on the time taken to import commits
which add many files.  Specifically, it makes them much faster for
pulls over the network because it avoids creating a new connection to
subversion for every added file.  In my testing, it dropped the time
taken to import a revision that adds ~6500 files from approximately 1
hour to 30 seconds.  I believe this test is representative of
real-world performance improvements.