svnwrap: allow overwriting directories in one revision
If replacing a directory at one path with a copy from somewhere else,
we need to first delete the old, and then create the new. Otherwise,
Subversion will bail with an error.

This could happen when moving a tag; in that case, we need to delete
the original tag and set the new one in the same revision.