push: obsolete rather than rebase & strip when enabled
This requires a few changes to wrappers.push() to use obsolescence
rather than strip and to make the rebase -- which is non-destructive
with obsolete active -- to no longer keep the originals. Possible
future work involves no longer relying on rebase for non-outgoing
revisions, and simply leaving them in the troubled state.

We test this feature by adding setting obsolete_mode_tests to True in
classes that push changes.