stupid: kill some dead and no longer needed code
While I was debugging an unrelated issue with stupid, some of the code
in branches_in_paths confused me, since it was using an undeclared
variable.  It looked like an attempt to short-circuit the
file/directory detection before talking to subversion that never quite
got finished.  The code is mostly unreachable, and obviously broken,
so I cleaned it up.  I also cleaned up some prepatory code that was
getting executed but appeared to only be useful for the
short-circuintg attempt.