stupid: cleanup unnecessary, always-true if statement
We had an unnecessary if statement in setupid.branches_in_paths() that
was checking a variable that was unconditionally set to True on the
previous line.  This was a remnant of a never-completed, and now
mostly-cleaned up attempt to short-circuit path type detection before
talking to subversion for some directories.

This removes the variable assignment and if statement, and moves the
body of the if up one level.