subdir: use util.dump and util.load for writing and reading
We need to change both svnmeta and svncommands at the same time since they are
heavily tied together.

The reason for this change is to remove the duplicate code for reading and
writing subdir present in svncommands.py. We will now use the standard
util.dump and util.load for writing and reading.

Due to the way json reads a string, the old format is still valid for use and
will be read correctly.