uuid: use util.dump and util.load for writing and reading
We need to change both svnmeta and svncommands at the same time since they are
heavily tied together.

The reason for this change is to remove the duplicate code for reading and
writing uuid present in svncommands.py. We will now use the standard util.dump
and util.load for writing and reading.

This presents a slight change in file format. Previously, the uuid file had the
format:

d073be05-634f-4543-b044-5fe20cf6d1d6[no newline]

and after this change, it is:

"d073be05-634f-4543-b044-5fe20cf6d1d6"[newline optional]

Due to the way json reads a string, the old format is still valid for use and
will be read correctly.