editor: fix replay handling for copied + modified symlinks
We strip the 'link ' prefix from symlinks when we store it in
Mercurial.  We reapply it when we start editing the file via
open_file, but not via add_file.  this means that modified symlniks
would replay correctly, but not copied and modified symlinks.  This
corrects that ommission.

--HG--
extra : rebase_source : 7adab26d6ec5ccdefa134061ee3ea454d690c099