svnmeta: explicitly test for None
There was a slight error in the way we were checking the value of 'c' via 'if
c'. If the type of 'c' is a bool then this could incorrectly be false. Instead,
we check for None explicitly. Nothing in production should have been affected
by this yet since this was only a problem with patches not yet released.