wrapper: delay marker creation in push
Before this patch we were creating obsolescence markers when pulling the
changesets from subversion. The counterpart stripping logic (when obsolescence
is not available) was done after pushing everything.

This patch moves the marker creation to the end of the push, to:
1) Make the code more readable (the marker creation and the stripping should
be in the same place)
2) Fix a behavior issue with inhibit(evolve) (see below)

With inhibit + hgsubversion enabled. When pushing to subversion we were
creating markers as expected but inhibiting the marker on the commit that was
the parent of the working copy making it therefore visible.
By delaying the marker creation to after the update we avoid this issue.