svnmeta: do not add instance references to the class object
This is part of a bigger plan to prevent a fd leak by making sure the sqlite
connection object is garbage collected.

Usually this is not a very serious issue but the testing framework will run
all tests in a single process by default. The fd leak will make tests fail
on systems with a low RLIMIT_NOFILE.

Previously, when we are adding new properties to the SVNMeta class, we bind
the current instance in the closure, which essentially prevents any SVNMeta
instances from being garbage collected (and its state like revmap). This
patch changed "self" from the closure one to the lambda argument to address
the issue.