wrappers: remove "svn.meta = meta" in pull
This is part of a bigger plan to prevent a fd leak because SqliteRevMap.db
never gets garbage collected.

The "svn.meta = meta" adds another reference to the SVNMeta object and is
never used. I changed it to "svn.meta = 1" and tests are passing so it looks
safe to remove.