maps: increase sqlite cache size automatically
Sqlite cache size is 2MB by default. Increasing the cache size can speed up
migration. In a large repo with millions of revisions, changing the cache
size to a big enough value reduces at least 30% of the migration time.

Give the general trends that sacrifices space for speed, this patch auto
tunes the cache size to be big enough to hold the entire database.