maps: disable iterating methods of RevMap
Since __iter__ is disabled in SqliteRevMap, we want to avoid the pattern
iterating RevMap as well. This patch disables all entries to keys, values and
items and makes it possible to bypass the restriction by setting "_allowiter".

Developers looking for a way to iterate the RevMap should add a new method
in RevMap and SqliteRevMap instead.

The rebuildmeta test needs iteration so we enable it explicitly there.