tests: optimise creating repositories and loading dumps
Previously, we'd use svnadmin for creating repositories and loading
dumps. That tends to be a bit slow, as it forks a new process and
loads the Subversion libraries into it. Instead, we extend our
existing Subversion wrappers and load the dumps using the API.

This is a noticable speedup. The only downside is that we rely on
Subversion and Subvertpy to correctly close all file descriptors; an
assumption which hasn't always held in the past.

I ran some benchmarks on my relatively slow Mac with $TMPDIR on a
ramdisk, and they showed a significant change:

I compared ten runs of each with Subvertpy:
     min:  -18.8% (299.1s -> 243.0s)
  median:  -20.0% (307.1s -> 245.6s)

...and three runs of each with SWIG:
     min:  -22.8% (368.7s -> 284.7s)
  median:  -25.7% (384.4s -> 285.5s)

(Since the timing measures wall clock time, the minimum time is likely
to be the most accurate and useful measurement.)