svnmeta: add a config option to use sqlite revmap
This patch adds a config option to choose the revmap implementation.

It's "plain" by default and can be set to "sqlite". The sqlite implementation
will migrate data from the plain revmap while the "plain" implementation will
complain if it sees the sqlite revmap.

A class variable is added for setting the default revmap implementation,
which is useful for testing.