maps: implement sqlite revmap
This patch adds the SqliteRevMap, which has a same interface with RevMap
but is backed by a sqlite database.

It uses database indexes to accelerate all kinds of queries and disables
iteration to prevent slow code being written in the future.

In practise, it should be faster on large repos with millions of svn
revisions but slower on small repos due to the overhead introduced.