tests: fix ipv6 support in test_push_command.py
This test runs the `svnserve` binary, which appears to resolve the hostname
passed as an argument using gethostbyname(3), which can only return an IPv4
address. When run on a system that is configured to only have an IPv6 address,
Mercurial will correctly resolve the hostname to an IPv6 address, but the SVN
server will be listening on an incorrect IPv4 address, causing the test to
fail. The workaround is to resolve the hostname using getaddrinfo(3) in the
test harness and pass the resulting address to svnserve.

--HG--
extra : rebase_source : afcad1de046cb0863696103c5eda1ce95d57bf1d