tests: when making a `memctx`, make sure to use a single repo instance
The way self.repo is dynamic produces bad lock behavior because the
`context.memctx` ends up with a different instance than self in
`localrepo.commitctx`, which means the callbacks in the `memctx` get
an unlocked repo instance. This causes lock warning failures. When
it's not a code freeze for core, we should probably:

 * Make lock failures hard, not just warnings
 * Stop holding a repo reference in memctx, or otherwise check it's the same
   instance as `self` during `localrepo.commitctx`

That's my best guess based on the (very hard to debug) test failures here.