test_util: make developer warnings cause our tests to fail
Previously hgsubversion didn't get any help from develwarn calls,
because we don't use core's testrunner. The upstream configitem
refactor was easier to hunt down using the develwarn being a failure,
which prompted this code. It also exposed several locking correctness
bugs, so it feels like it's worth keeping around long term.

We don't bother with develwarn on anything so old that we don't try to
use smartset. The smartset deprecation was not handled gracefully, so
hg complains for many versions without actually breaking anything, and
it'd pointlessly complicate our code to handle smartset living in
multiple places between when the warning was introduced and the
present.