verify: use ui.write() instead of ui.warn()
Looks like ui.warn() is being written to output directly and it's not pushed to
buffer which ui objects maintain. Let's use ui.write() instead which is used in
the other instance of same error messages too.

This fixes test_utility_commands.py which was failing for me on tip of default
branch.